cut/paste test stubs for header tests
